Nft metadata

+++ contracts/AaveStrategy.sol
@@ -7,15 +7,18 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IPool} from "@aave/core-v3/contracts/interfaces/IPool.sol";
 
 contract AaveStrategy {
-
     /**
      * @param pool AAVE pool proxy address
      * @param token token address
      * @param amount amount to deposit
-     * 
+     *
      * @dev Sender should first approve this contract to spend the amount of tokens to be deposited
      */
-    function deposit(address pool, address token, uint256 amount) external {
+    function deposit(
+        address pool,
+        address token,
+        uint256 amount
+    ) external {
         // Send tokens to this contract
         IERC20(token).transferFrom(msg.sender, address(this), amount);
 
@@ -34,7 +37,12 @@ contract AaveStrategy {
      * @param amount amount to withdraw
      * @param user address of the user who should receive the withdrawn tokens
      */
-    function claim(address pool, address token, uint256 amount, address user) external {
+    function claim(
+        address pool,
+        address token,
+        uint256 amount,
+        address user
+    ) external {
         // Withdraw from AAVE to the user
         IPool(pool).withdraw(token, amount, user);
     }
@@ -44,8 +52,12 @@ contract AaveStrategy {
      * @param token token address
      * @param user address of the user who should receive the withdrawn tokens
      */
-    function claimAll(address pool, address token, address user) external {
+    function claimAll(
+        address pool,
+        address token,
+        address user
+    ) external {
         // Withdraw from AAVE to the user
-        IPool(pool).withdraw(token, type(uint).max, user);
+        IPool(pool).withdraw(token, type(uint256).max, user);
     }
 }
